requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.3.1
cssselect2==0.4.1
cycler==0.10.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.42.0
django-bootstrap-static==4.2.1
django-countries==6.1.3
django-crispy-forms==1.9.2
django-filter==2.3.0
django-heroku==0.3.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
python-barcode==0.13.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
qrcode==6.1
reportlab==3.5.57
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tinycss2==1.1.0
urllib3==1.26.2
webencodings==0.5.1
whitenoise==5.2.0

